Refactored Tracks__List
